CA-455 (code review changes) allow moderator to mute another participant
